Add printf attribute to ConversionError
This exposes some useful warnings around wrong argument types supplied
to `ConversionError`, including a crash in `DecodeTables`.

+++ src/conv/graphic.c
@@ -330,7 +330,7 @@ static bool ValidatePatch(const struct patch_header *hdr, const uint8_t *srcbuf,
 		if (off > srcbuf_len - 1) {
 			ConversionError("Corrupted patch: column %d has "
 			                "invalid offset %d > %d",
-			                x, off, srcbuf_len - 1);
+			                x, off, (int) srcbuf_len - 1);
 			return false;
 		}
 		while (srcbuf[off] != 0xff) {
